Use EdgeDetection to render tiles in example
- Complement example cross shape with a box shape
- List all palette tiles by hand
- Add pipe and end tiles to definition
- Render tiles using palette next to existing views

Thoughts:
The palette defenition is quite verbose and result is fairly ugly.
It's clear that the 'EdgeDetection' module needs a bit of pondering
and a few more iterations before it's included in the public API.

Perhappes it's not even a v1.0 feature,
or maybe it's a problem big enough for a separate package.
Thinkering does however illuminate more basic features that
this package is currently lacking.